Fix crash when `is fuzzed` subs of same name
Since subs are lexical, it's possible to add more than one
sub with the same name. This causes a crash because hash is
type-constrainted to Routine, but Hash.push with existing key
creates an Array there.

Fix by removing that type constraint and modifying run-tests routine
to be able to handle multi-named subs in a key.